seperated cli command generator from curl command generator
