+++ packages/manager/src/utilities/generate-cURL.test.ts
@@ -1,4 +1,4 @@
-import generateCurlCommand, { generateCLICommand } from './generate-cURL';
+import generateCurlCommand from './generate-cURL';
 
 const linodeData = {
   label: 'linode-1',
@@ -6,15 +6,6 @@ const linodeData = {
   region: 'us-east',
   type: 'g6-standard-2',
 };
-
-const linodeDataForCLI = `
---label linode-1
---root_pass aComplex@Password
---region us-east
---type g6-standard-2
-`
-  .trimStart()
-  .trimEnd();
 
 const linodeDataString = JSON.stringify(linodeData);
 
@@ -52,24 +43,5 @@ describe('generateCurlCommand', () => {
   it('should return a curl command that has a data option set to the argument passed to it', () => {
     const generatedCommand = generateCurlCommand(linodeData);
     expect(generatedCommand).toMatch(`-d ${linodeDataString}`);
-  });
-});
-
-describe('generateCLICommand', () => {
-  it('should return a linode-cli command', () => {
-    const generatedCommand = generateCLICommand(linodeData);
-    expect(generatedCommand.startsWith('linode-cli')).toBeTruthy();
-  });
-
-  it('should return a linode-cli linodes create command', () => {
-    const generatedCommand = generateCLICommand(linodeData);
-    expect(
-      generatedCommand.startsWith('linode-cli linodes create')
-    ).toBeTruthy();
-  });
-
-  it('should return a linode-cli command with the data provided formatted as arguments', () => {
-    const generatedCommand = generateCLICommand(linodeData);
-    expect(generatedCommand).toMatch(linodeDataForCLI);
   });
 });
